Replace ushort/ulong with uint16_t/uint32_t
All the documentation I find says that these COFF fields are of type
"unsigned short" and "unsigned long".  It may be 16/32 bits, or it may
be machine-dependent.  Byte ordering is probably machine-dependent.
Let's assume 16/32 bit for now since any existing COFF binaries probably
aren't 64-bit.

+++ loader/IntelCoffFile.h
@@ -3,18 +3,19 @@
 
 #include "BinaryFile.h"
 #include "SymTab.h"
+#include <stdint.h>
 
 #define PACKED __attribute__((packed))
 
 struct coff_header
 {
-        ushort  coff_magic;
-        ushort  coff_sections;
-        ulong   coff_timestamp;
-        ulong   coff_symtab_ofs;
-        ulong   coff_num_syment;
-        ushort  coff_opthead_size;
-        ushort  coff_flags;
+        uint16_t coff_magic;
+        uint16_t coff_sections;
+        uint32_t coff_timestamp;
+        uint32_t coff_symtab_ofs;
+        uint32_t coff_num_syment;
+        uint16_t coff_opthead_size;
+        uint16_t coff_flags;
 } PACKED;
 
 class IntelCoffFile : public BinaryFile
